+++ oap-server/analyzer/meter-analyzer/src/main/java/org/apache/skywalking/oap/meter/analyzer/dsl/DSL.java
@@ -25,6 +25,8 @@ import groovy.util.DelegatingScript;
 import java.lang.reflect.Array;
 import java.util.List;
 import java.util.Map;
+
+import org.apache.skywalking.oap.meter.analyzer.dsl.registry.ProcessRegistry;
 import org.apache.skywalking.oap.meter.analyzer.dsl.tagOpt.K8sRetagType;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.core.source.DetectPoint;
@@ -54,6 +56,7 @@ public final class DSL {
         icz.addImport("K8sRetagType", K8sRetagType.class.getName());
         icz.addImport("DetectPoint", DetectPoint.class.getName());
         icz.addImport("Layer", Layer.class.getName());
+        icz.addImport("ProcessRegistry", ProcessRegistry.class.getName());
         cc.addCompilationCustomizers(icz);
 
         final SecureASTCustomizer secureASTCustomizer = new SecureASTCustomizer();
@@ -73,7 +76,8 @@ public final class DSL {
                          .add(K8sRetagType.class)
                          .add(DetectPoint.class)
                          .add(Layer.class)
-                         .build());
+                         .add(ProcessRegistry.class)
+                .build());
         cc.addCompilationCustomizers(secureASTCustomizer);
 
         GroovyShell sh = new GroovyShell(new Binding(), cc);